rf: move `check_source_jsonschema_validation_error`
From `cli/models.py` to `models.py`
