+++ src/dandisets_linkml_status_tools/cli/models.py
@@ -3,10 +3,13 @@ from datetime import datetime
 from typing import Annotated, Any, NamedTuple
 
 from dandi.dandiapi import VersionStatus
-from jsonschema.exceptions import ValidationError
 from linkml.validator.report import ValidationResult
 from pydantic import AfterValidator, BaseModel, Json, PlainSerializer, TypeAdapter
 from typing_extensions import TypedDict  # Required for Python < 3.12 by Pydantic
+
+from dandisets_linkml_status_tools.models import (
+    check_source_jsonschema_validation_error,
+)
 
 
 class JsonValidationErrorView(BaseModel):
@@ -34,32 +37,6 @@ PolishedValidationResult = TypedDict(
     "PolishedValidationResult",
     field_annotations,
 )
-
-
-def check_source_jsonschema_validation_error(
-    results: list[ValidationResult],
-) -> list[ValidationResult]:
-    """
-    Check if the `source` field of each `ValidationResult` object in a given list is a
-    `jsonschema.exceptions.ValidationError` object.
-
-    :param results: The list of `ValidationResult` objects to be checked.
-
-    :return: The list of `ValidationResult` objects if all `source` fields are
-        `jsonschema.exceptions.ValidationError` objects.
-
-    :raises ValueError: If the `source` field of a `ValidationResult` object is not a
-        `jsonschema.exceptions.ValidationError` object.
-    """
-    for result in results:
-        result_source = result.source
-        if not isinstance(result_source, ValidationError):
-            msg = (
-                f"Expected `source` field of a `ValidationResult` object to be "
-                f"a {ValidationError!r} object, but got {result_source!r}"
-            )
-            raise ValueError(msg)  # noqa: TRY004
-    return results
 
 
 def polish_validation_results(
