+++ src/dandisets_linkml_status_tools/models.py
@@ -1,5 +1,7 @@
 from typing import Any
 
+from jsonschema import ValidationError
+from linkml.validator.report import ValidationResult
 from pydantic import BaseModel, Json
 
 PydanticValidationErrsType = list[dict[str, Any]]
@@ -35,3 +37,29 @@ class AssetValidationReport(ValidationReport):
 
     asset_id: str | None
     asset_path: str | None
+
+
+def check_source_jsonschema_validation_error(
+    results: list[ValidationResult],
+) -> list[ValidationResult]:
+    """
+    Check if the `source` field of each `ValidationResult` object in a given list is a
+    `jsonschema.exceptions.ValidationError` object.
+
+    :param results: The list of `ValidationResult` objects to be checked.
+
+    :return: The list of `ValidationResult` objects if all `source` fields are
+        `jsonschema.exceptions.ValidationError` objects.
+
+    :raises ValueError: If the `source` field of a `ValidationResult` object is not a
+        `jsonschema.exceptions.ValidationError` object.
+    """
+    for result in results:
+        result_source = result.source
+        if not isinstance(result_source, ValidationError):
+            msg = (
+                f"Expected `source` field of a `ValidationResult` object to be "
+                f"a {ValidationError!r} object, but got {result_source!r}"
+            )
+            raise ValueError(msg)  # noqa: TRY004
+    return results
